requirements.txt
Flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
python-socketio==5.10.0
python-engineio==4.8.0
eventlet==0.33.3
